Keep warm capacity helpers Kubernetes-only
