+++ controlplane/warm_capacity_targets.go
@@ -15,31 +15,6 @@ type dynamicWarmCapacityConfig struct {
 	MaxWorkers          int
 }
 
-func dynamicWarmCapacityConfigFromK8s(cfg K8sConfig) dynamicWarmCapacityConfig {
-	return dynamicWarmCapacityConfig{
-		Enabled:             cfg.DynamicWarmCapacityEnabled,
-		MissesPerWorker:     cfg.WarmCapacityMissesPerWorker,
-		DynamicImageCeiling: cfg.WarmCapacityDynamicImageCeiling,
-		DynamicTotalCeiling: cfg.WarmCapacityDynamicTotalCeiling,
-		MaxWorkers:          cfg.MaxWorkers,
-	}
-}
-
-func computeBaseWarmCapacityTargets(defaultImage string, sharedWarmTarget int, perImageFloors map[string]int) map[string]int {
-	out := sanitizeWarmCapacityTargets(perImageFloors)
-	defaultImage = strings.TrimSpace(defaultImage)
-	if defaultImage == "" {
-		return out
-	}
-	if sharedWarmTarget < 0 {
-		sharedWarmTarget = 0
-	}
-	if sharedWarmTarget > out[defaultImage] {
-		out[defaultImage] = sharedWarmTarget
-	}
-	return out
-}
-
 func computeDynamicWarmCapacityTargets(baseTargets map[string]int, aggregates []configstore.WarmCapacityMissAggregate, cfg dynamicWarmCapacityConfig) map[string]int {
 	targets := sanitizeWarmCapacityTargets(baseTargets)
 	if !cfg.Enabled {
